refactor: move caps to caps.rs inside caps module
